Fixed the Gamestop price and also changed the ProcessCheckoutInfo struct

+++ juiced.sitescripts/bestbuy/task.go
@@ -1112,7 +1112,7 @@ func (task *Task) PlaceOrder(startTime time.Time) (bool, enums.OrderStatus) {
 		ItemName:     task.CheckoutInfo.ItemName,
 		Sku:          task.CheckoutInfo.SKUInStock,
 		Retailer:     enums.BestBuy,
-		Price:        task.CheckoutInfo.Price,
+		Price:        float64(task.CheckoutInfo.Price),
 		Quantity:     1,
 		MsToCheckout: time.Since(startTime).Milliseconds(),
 	})
